requirements.txt
# Annam AI - Agricultural Intelligence System
# Python dependencies for the complete system (Python 3.11+)

# Core ML/DL frameworks
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP and transformers
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.14.0

# Vector databases and RAG
faiss-cpu>=1.7.0
chromadb>=0.4.0

# Web frameworks
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
gradio>=3.40.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
joblib>=1.3.0
python-dotenv>=1.0.0
plotly>=5.15.0

# Mobile export
onnx>=1.14.0
onnxruntime>=1.15.0

# Additional ML tools
optuna>=3.3.0
mlflow>=2.5.0
statsmodels>=0.14.0
albumentations>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# NOTE:
# sqlite3 is part of Python's standard library — no need to install via pip.
